services/backend/pyproject.toml
[tool.poetry]
name = "quality_assesment"
version = "0.1.0"
description = ""
authors = ["Andrew281213 <andrey.domakhin8@ya.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
aerich = "^0.7.1"
uvicorn = "^0.21.1"
python-dotenv = "^1.0.0"
tortoise-orm = "^0.19.3"
asyncpg = "^0.27.0"

[tool.poetry.dev-dependencies]


[tool.aerich]
tortoise_orm = "src.database.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
